code/lsp: split request helpers into submodule
`src/code/lsp.rs` had grown past 1200 lines as workspace_symbol /
type_definition / implementation / call_hierarchy landed under
YYC-44. Convert the file into a `code/lsp/` module directory:
mod.rs keeps the LspServer + LspManager + RPC framing core, and a
new `requests.rs` submodule holds the high-level helpers
(goto_definition, find_references, hover, type_definition,
implementation, prepare_call_hierarchy + incoming/outgoing,
workspace_symbol, diagnostics_for, prepare_request,
flatten_goto_response).

Mechanical move only — every helper keeps its public signature so
the `crate::code::lsp::goto_definition` import path still works.
mod.rs re-exports `requests::*`. `path_to_uri` and the readiness
constant lift to `pub(super)` so the submodule can use them.
The two test sites still pass against the same module name.
